Empty commit at end of history preceding clang-format style transition
This is an empty commit that precedes an automatic application of
clang-format to update the C++ style of our entire source tree.
This may be helpful to rebase a topic branch that was originally
based on a commit preceding the transition.  One may first rebase
the topic on this commit.  Then use one of the following approaches.

*   Rewrite the topic, including this commit, using `git filter-branch`
    `--tree-filter` with `Utilities/Scripts/clang-format.bash` to update
    the style in every commit.  Rebase the revised topic, excluding the
    rewrite of this commit, on the style transition commit.

OR

*   Add a `.git/info/grafts` entry to change the parent of the first
    commit in the topic from this commit to the style transition commit.
    Rewrite the topic using `git filter-branch --tree-filter` with
    `Utilities/Scripts/clang-format.bash` to update the style in every
    commit.  Then remove the graft, which was resolved by the filter.

See `git help filter-branch` and `git help repository-layout` for
details.