Support for many custom commands in Windows.
Windows has a limit of 8191 characters per command line, so this patch
creates a batch file instead if the commands get too long.

don't be noisy in the custom command, @echo off

Uses a hash of the commands as the filename for a .cmd file in case
command line is too long. Also, this has a boolean parameter to check if
no ninja variables are set. Otherwise we can't generate the file.